+++ plugin/src/main/java/com/wkl/manifest/config/AbsConfig.java
@@ -1,11 +1,18 @@
 package com.wkl.manifest.config;
 
-import com.wkl.manifest.iinterface.ParseProperty;
-
 /**
  * Created by <a href="mailto:wangkunlin1992@gmail.com">Wang kunlin</a>
  * <p>
  * On 2018-05-10
  */
-abstract class AbsConfig implements ParseProperty {
+abstract class AbsConfig {
+    public boolean mRemoved;
+
+    public void remove(boolean remove) {
+        mRemoved = remove;
+    }
+
+    void parseProperty(StringBuilder container) {
+        container.append(mRemoved);
+    }
 }