+++ plugin/src/main/java/com/wkl/manifest/config/ActivityConfig.java
@@ -1,24 +1,15 @@
 package com.wkl.manifest.config;
 
+import org.gradle.api.Project;
+
 /**
  * Created by <a href="mailto:wangkunlin1992@gmail.com">Wang kunlin</a>
  * <p>
  * On 2018-04-16
  */
-public class ActivityConfig extends AbsConfig {
-    public String name;
-    public boolean removed;
+public class ActivityConfig extends ComponentConfig {
 
-    public void remove(boolean remove) {
-        removed = remove;
-    }
-
-    public void name(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public void parseProperty(StringBuilder container) {
-        container.append("ActivityConfig:").append(name).append(removed);
+    ActivityConfig(Project project) {
+        super(project);
     }
 }