+++ plugin/src/main/java/com/wkl/manifest/config/ApplicationConfig.java
@@ -1,64 +1,38 @@
 package com.wkl.manifest.config;
 
+import org.gradle.api.Project;
+
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Map;
-import java.util.Set;
+
+import groovy.lang.Closure;
 
 /**
  * Created by <a href="mailto:wangkunlin1992@gmail.com">Wang kunlin</a>
  * <p>
  * On 2018-04-13
  */
-public class ApplicationConfig extends AbsConfig {
+public class ApplicationConfig extends AbsAppConfig {
 
-    private Map<String, String> mToAddAttrs = new HashMap<>();
-    private Map<String, String> mToModAttrs = new HashMap<>();
-    private Set<String> mToDelAttrs = new HashSet<>();
+    private Map<String, ActivityConfig> mActivityConfigs = new HashMap<>();
 
-    public void addAttr(String name, String value) {
-        mToAddAttrs.put(name, value);
+    public ApplicationConfig(Project project) {
+        super(project);
     }
 
-    public void delAttr(String name) {
-        mToDelAttrs.add(name);
+    public Map<String, ActivityConfig> getActivityConfigs() {
+        return mActivityConfigs;
     }
 
-    public void modAttr(String name, String newValue) {
-        mToModAttrs.put(name, newValue);
-    }
-
-    public Map<String, String> getToAddAttrs() {
-        return mToAddAttrs;
-    }
-
-    public Map<String, String> getToModAttrs() {
-        return mToModAttrs;
-    }
-
-    public Set<String> getToDelAttrs() {
-        return mToDelAttrs;
+    public void activity(Closure closure) {
+        ActivityConfig config = new ActivityConfig(mProject);
+        mProject.configure(config, closure);
+        mActivityConfigs.put(config.mName, config);
     }
 
     @Override
     public void parseProperty(StringBuilder container) {
-        container.append("ApplicationConfig:");
-        fillProperty(container, mToAddAttrs, "ToAddAttrs");
-        fillProperty(container, mToModAttrs, "ToModAttrs");
-        fillProperty(container, mToDelAttrs, "ToAddAttrs");
-    }
-
-    private void fillProperty(StringBuilder sb, Map<String, String> attr, String name) {
-        sb.append(name).append("%");
-        attr.forEach((k, v) -> sb.append(k).append('-').append(v).append("#"));
-        sb.deleteCharAt(sb.length() - 1);
-        sb.append("@");
-    }
-
-    private void fillProperty(StringBuilder sb, Set<String> attr, String name) {
-        sb.append(name).append("%");
-        attr.forEach(v -> sb.append(v).append('-').append("#"));
-        sb.deleteCharAt(sb.length() - 1);
-        sb.append("@");
+        super.parseProperty(container);
+        mActivityConfigs.forEach((s, activityConfig) -> activityConfig.parseProperty(container));
     }
 }