+++ plugin/src/main/java/com/wkl/manifest/plugin/ManifestExtension.java
@@ -1,16 +1,11 @@
 package com.wkl.manifest.plugin;
 
 
-import com.wkl.manifest.config.ActivityConfig;
 import com.wkl.manifest.config.ApplicationConfig;
+import com.wkl.manifest.utils.Utils;
 
 import org.gradle.api.Action;
 import org.gradle.api.Project;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import groovy.lang.Closure;
 
 /**
  * Created by <a href="mailto:wangkunlin1992@gmail.com">Wang kunlin</a>
@@ -20,12 +15,9 @@ import groovy.lang.Closure;
 public class ManifestExtension {
 
     private ApplicationConfig mApplicationConfig;
-    private List<ActivityConfig> mActivityConfigs = new ArrayList<>();
-    private Project mProject;
 
     public ManifestExtension(Project project) {
-        mApplicationConfig = new ApplicationConfig();
-        mProject = project;
+        mApplicationConfig = new ApplicationConfig(project);
     }
 
     public void application(Action<ApplicationConfig> action) {
@@ -40,20 +32,9 @@ public class ManifestExtension {
         return project.getExtensions().getByType(ManifestExtension.class);
     }
 
-    public List<ActivityConfig> getActivityConfigs() {
-        return mActivityConfigs;
-    }
-
-    public void activity(Closure closure) {
-        ActivityConfig config = new ActivityConfig();
-        mProject.configure(config, closure);
-        mActivityConfigs.add(config);
-    }
-
     public String parseProperty() {
         StringBuilder container = new StringBuilder("ManifestExtension:");
         mApplicationConfig.parseProperty(container);
-        mActivityConfigs.forEach(activityConfig -> activityConfig.parseProperty(container));
-        return container.toString();
+        return Utils.generateMD5(container.toString());
     }
 }